third_party/WebKit/LayoutTests/fullscreen/enter-exit-full-screen-hover-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.webkitIsFullScreen is true
PASS getHoverActiveState(enterButton) is "default"
PASS document.webkitIsFullScreen is false
PASS getHoverActiveState(exitButton) is "default"
Go full screen Exit full screen
EVENT(webkitfullscreenchange)
EVENT(webkitfullscreenchange)
END OF TEST
